requirements.txt
pinecone
google-generativeai
pymupdf 
pillow 
duckduckgo-search 
python-dotenv 
"camelot-py[cv]"
Pillow